FifthMilestone/ws/requirements.txt
Flask
Flask-HAL
PyJWT 
gunicorn
python-dotenv
redis
six
python-jose
pika